+++ src/lib/ast/transform/compute.ts
@@ -33,10 +33,44 @@ export function compute(transformLists: Token[]): {
     let matrix: Matrix | null = identity();
     let mat: Matrix;
     let transforms: Token[];
+    let shouldComputeMatrix: boolean = true;
     const cumulative: Token[] = [];
 
+    // all transform function arguments must be 0 or scale(1)
+    for (const transform of transformLists) {
+        switch ((transform as FunctionToken).val) {
+            case "rotate":
+            case "rotateX":
+            case "rotateY":
+            case "rotateZ":
+            case "rotate3d":
+                for (const child of (transform as FunctionToken).chi) {
+                    if (child.typ == EnumToken.WhitespaceTokenType || child.typ == EnumToken.CommaTokenType) {
+                        continue;
+                    }
+
+                    if (
+                        (child.typ != EnumToken.AngleTokenType &&
+                            child.typ != EnumToken.NumberTokenType &&
+                            child.typ != EnumToken.PercentageTokenType) ||
+                        // angle >= 360deg, do not transform
+                        Math.abs(getAngle(child as NumberToken)) >= 1
+                    ) {
+                        shouldComputeMatrix = false;
+                        null;
+                    }
+                }
+
+                break;
+        }
+
+        if (!shouldComputeMatrix) {
+            break;
+        }
+    }
+
     for (const transformList of splitTransformList(transformLists)) {
-        mat = computeMatrix(transformList, identity()) as Matrix;
+        mat = shouldComputeMatrix ? (computeMatrix(transformList, identity()) as Matrix) : null;
 
         if (mat == null) {
             return null;
@@ -73,72 +107,6 @@ export function compute(transformLists: Token[]): {
         cumulative,
         minified: minify(matrix) ?? [serialized],
     };
-
-    // valid identity matrix
-    if (
-        (result.minified.length == 1 &&
-            result.minified[0].typ == EnumToken.IdenTokenType &&
-            (result.minified[0] as IdentToken).val == "none") ||
-        (result.cumulative.length == 1 &&
-            result.cumulative[0].typ == EnumToken.IdenTokenType &&
-            (result.cumulative[0] as IdentToken).val == "none") ||
-        (result.matrix?.typ == EnumToken.IdenTokenType && (result.matrix as IdentToken).val == "none")
-    ) {
-        // all transform function arguments must be 0 or scale(1)
-        for (const transform of transformLists) {
-            switch ((transform as FunctionToken).val) {
-                case "translate":
-                case "translateX":
-                case "translateY":
-                case "translateZ":
-                case "translate3d":
-                case "rotate":
-                case "rotateX":
-                case "rotateY":
-                case "rotateZ":
-                case "rotate3d":
-                case "skew":
-                case "skewX":
-                case "skewY":
-                    for (const child of (transform as FunctionToken).chi) {
-                        if (child.typ == EnumToken.WhitespaceTokenType || child.typ == EnumToken.CommaTokenType) {
-                            continue;
-                        }
-
-                        if (
-                            (child.typ != EnumToken.AngleTokenType &&
-                                child.typ != EnumToken.NumberTokenType &&
-                                child.typ != EnumToken.PercentageTokenType) ||
-                            getNumber(child as NumberToken) != 0
-                        ) {
-                            return null;
-                        }
-                    }
-
-                    break;
-
-                case "scale":
-                case "scaleX":
-                case "scaleY":
-                case "scaleZ":
-                case "scale3d":
-                    for (const child of (transform as FunctionToken).chi) {
-                        if (child.typ == EnumToken.WhitespaceTokenType || child.typ == EnumToken.CommaTokenType) {
-                            continue;
-                        }
-
-                        if (
-                            (child.typ != EnumToken.NumberTokenType && child.typ != EnumToken.PercentageTokenType) ||
-                            getNumber(child as NumberToken) != 1
-                        ) {
-                            return null;
-                        }
-                    }
-
-                    break;
-            }
-        }
-    }
 
     return result;
 }