+++ src/lib/ast/transform/minify.ts
@@ -301,7 +301,7 @@ export function eqMatrix(a: FunctionToken | Matrix, b: Token[]): boolean {
 }
 
 export function minifyTransformFunctions(transform: FunctionToken): FunctionToken {
-    const name: string = transform.val.toLowerCase();
+    const name: string = transform.val;
 
     if ("skewX" == name) {
         transform.val = "skew";