+++ src/lib/parser/tokenize.ts
@@ -401,7 +401,7 @@ export class Tokenizer {
      * @param parseInfo
      * @returns
      */
-    consumeURLToken(parseInfo: ParseInfo): this {
+    consumeStringAsURLToken(parseInfo: ParseInfo): this {
         const quote: number = this.advance(parseInfo).charCodeAt(0);
         let charCode: number;
         let decodeSegments: boolean = false;
@@ -959,7 +959,7 @@ export class Tokenizer {
         return 0;
     }
 
-    parseURLToken(parseInfo: ParseInfo, endPosition: number): this {
+    consumeURLToken(parseInfo: ParseInfo, endPosition: number): this {
         let charCode: number;
 
         // consume an <url>
@@ -970,7 +970,7 @@ export class Tokenizer {
         charCode = this.peekCharCode(parseInfo);
 
         if (charCode == TokenMap.DOUBLE_QUOTE || charCode == TokenMap.SINGLE_QUOTE) {
-            return this.consumeURLToken(parseInfo);
+            return this.consumeStringAsURLToken(parseInfo);
         }
 
         do {
@@ -1285,25 +1285,25 @@ export class Tokenizer {
      * @param end
      * @returns
      */
-    isIdentToken(parseInfo: ParseInfo, start?: number, end?: number): boolean {
+    isIdentToken(parseInfo: ParseInfo /* , start?: number, end?: number */): boolean {
         let j: number = parseInfo.currentPosition - parseInfo.offset;
         let i: number = parseInfo.position - parseInfo.offset;
 
-        if (start != null) {
-            if (end == null) {
-                if (start < 0) {
-                    j += start;
-                } else {
-                    i += start;
-                }
-            } else {
-                if (end < 0) {
-                    j += end;
-                } else {
-                    j = parseInfo.position + end;
-                }
-            }
-        }
+        // if (start != null) {
+        //     if (end == null) {
+        //         if (start < 0) {
+        //             j += start;
+        //         } else {
+        //             i += start;
+        //         }
+        //     } else {
+        //         if (end < 0) {
+        //             j += end;
+        //         } else {
+        //             j = parseInfo.position + end;
+        //         }
+        //     }
+        // }
 
         j--;
 
@@ -1361,18 +1361,18 @@ export class Tokenizer {
      * @param parseInfo
      * @returns
      */
-    isPseudo(parseInfo: ParseInfo): boolean {
-        let position: number = parseInfo.currentPosition - parseInfo.offset;
-        let endPosition: number = parseInfo.currentPosition - parseInfo.offset;
-        return (parseInfo.stream.charAt(position) == ":" &&
-            parseInfo.stream.charAt(endPosition - 1) == "(" &&
-            (parseInfo.stream.charAt(position + 1) == ":"
-                ? this.isIdentToken(parseInfo, 2, -1)
-                : this.isIdentToken(parseInfo, 1, -1))) ||
-            parseInfo.stream.charAt(position + 1) == ":"
-            ? this.isIdentToken(parseInfo, 2)
-            : this.isIdentToken(parseInfo, 1);
-    }
+    // isPseudo(parseInfo: ParseInfo): boolean {
+    //     let position: number = parseInfo.currentPosition - parseInfo.offset;
+    //     let endPosition: number = parseInfo.currentPosition - parseInfo.offset;
+    //     return (parseInfo.stream.charAt(position) == ":" &&
+    //         parseInfo.stream.charAt(endPosition - 1) == "(" &&
+    //         (parseInfo.stream.charAt(position + 1) == ":"
+    //             ? this.isIdentToken(parseInfo, 2, -1)
+    //             : this.isIdentToken(parseInfo, 1, -1))) ||
+    //         parseInfo.stream.charAt(position + 1) == ":"
+    //         ? this.isIdentToken(parseInfo, 2)
+    //         : this.isIdentToken(parseInfo, 1);
+    // }
 
     /**
      *
@@ -1464,8 +1464,7 @@ export class Tokenizer {
         while ((charCode = this.peekCharCode(parseInfo)) == charCode) {
             if (this.state === EnumToken.UrlFunctionTokenDefType) {
                 this.state = null;
-                return this.parseURLToken(parseInfo, endPosition);
-                continue;
+                return this.consumeURLToken(parseInfo, endPosition);
             }
 
             if (parseInfo.position == parseInfo.currentPosition) {
@@ -1638,15 +1637,15 @@ export class Tokenizer {
                 // '('
                 case TokenMap.LEFT_PARENTHESIS:
                     if (parseInfo.position < parseInfo.currentPosition) {
-                        if (
-                            parseInfo.stream[parseInfo.position - parseInfo.offset] === ":" &&
-                            this.isPseudo(parseInfo)
-                        ) {
-                            this.advance(parseInfo);
-                            return this.makeToken(parseInfo, EnumToken.PseudoClassFunctionTokenDefType);
+                        // if (
+                        //     parseInfo.stream[parseInfo.position - parseInfo.offset] === ":" &&
+                        //     this.isPseudo(parseInfo)
+                        // ) {
+                        //     this.advance(parseInfo);
+                        //     return this.makeToken(parseInfo, EnumToken.PseudoClassFunctionTokenDefType);
 
-                            break;
-                        } else if (this.isIdentToken(parseInfo)) {
+                        // } else 
+                            if (this.isIdentToken(parseInfo)) {
                             const hint: EnumToken = this.startsWith(parseInfo, "--")
                                 ? EnumToken.CustomFunctionTokenDefType
                                 : (getSymbolHint(