+++ CHANGELOG.md
@@ -2,8 +2,28 @@
 
 # v1.6.1
 
-- [x] fix bugs when computing path functions.
+- [x] fix bugs when computing math functions.
 - [x] reduce array mutations
+- [x] do not minify transform function when the absolute angle is greater or equals to 360deg
+
+```css
+
+  a:hover {
+
+transform: rotate(calc(2.5*pi * 1rad))  }
+  .now {
+    transform: translate(100px, 100px) rotate(1215deg) skewX(10deg);
+}
+```
+should be minified to
+```css
+a:hover {
+ transform: rotate(450deg)
+}
+.now {
+ transform: translate(100px,100px)rotate(1215deg)skew(10deg)
+}
+```
 
 # v1.6.0