+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [x] fix bugs when computing math functions.
 - [x] reduce array mutations
 - [x] do not minify transform function when the absolute angle is greater or equals to 360deg
-- [x] spporting additional shorthands: scroll-timeline, scroll-padding, scroll-margin.
+- [x] supporting additional shorthands: scroll-timeline, scroll-padding, scroll-margin.
 
 ```css